refactor(app): useModuleCalibrationStatus
