+++ app/src/organisms/Devices/hooks/index.ts
@@ -6,7 +6,6 @@ export * from './useIsRobotViewable'
 export * from './useLights'
 export * from './useLEDLights'
 export * from './useLPCSuccessToast'
-export * from './useModuleCalibrationStatus'
 export * from './usePipetteOffsetCalibrations'
 export * from './usePipetteOffsetCalibration'
 export * from './useTipLengthCalibrations'
