+++ app/src/resources/runs/index.ts
@@ -25,3 +25,4 @@ export * from './useRunCreatedAtTimestamp'
 export * from './useLPCDisabledReason'
 export * from './useUnmatchedModulesForProtocol'
 export * from './useModuleRenderInfoForProtocolById'
+export * from './useModuleCalibrationStatus'
